content/media/mediasource/test/test_SplitAppend.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: append initialization and media segment separately</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(function () {
  var ms = new MediaSource();

  var v = document.createElement("video");
  v.preload = "auto";
  v.src = URL.createObjectURL(ms);
  document.body.appendChild(v);

  ms.addEventListener("sourceopen", function () {
    var sb = ms.addSourceBuffer("video/webm");

    fetchWithXHR("seek.webm", function (arrayBuffer) {
      sb.appendBuffer(new Uint8Array(arrayBuffer, 0, 318));
      sb.appendBuffer(new Uint8Array(arrayBuffer, 318));
      ms.endOfStream();
      v.play();
    });
  });

  v.addEventListener("ended", function () {
    is(v.duration, 4, "Video has correct duration");
    is(v.currentTime, 4, "Video has played to end");
    v.parentNode.removeChild(v);
    SimpleTest.finish();
  });
});

</script>
</pre>
</body>
</html>
